+++ hhset.c
@@ -2,11 +2,11 @@
 #include "hhash.h"
 #include "hhset.h"
 
-static uint id(uint x) { return x - 1; }
+static ulong id(ulong x) { return x - 1; }
 
-static int neq(uint x, uint y) { return x != y; }
+static int neq(ulong x, ulong y) { return x != y; }
 
-HHSet *hhsetnew(uint n, uint (*hash)(uint), int (*cmp)(uint, uint)) {
+HHSet *hhsetnew(ulong n, ulong (*hash)(ulong), int (*cmp)(ulong, ulong)) {
   HHSet *S = malloc(sizeof(HHSet));
   S->T = hhashnew(n);
   S->hash = (hash != NULL) ? hash : id;
@@ -21,19 +21,19 @@ void hhsetfree(HHSet *S) {
 
 int hhsetcopy(HHSet *S, HHSet *T) {
   for (int i = 0; i < S->T->n; ++i) {
-    uint x = S->T->V[i];
+    ulong x = S->T->V[i];
     if (x) {
-      uint h = T->hash(x);
+      ulong h = T->hash(x);
       if (!hhashput(T->T, h, x)) return 0;
     }
   }
   return 1;
 }
 
-static uint hunt(HHSet *S, uint h, uint k, int kill) {
+static ulong hunt(HHSet *S, ulong h, ulong k, int kill) {
   int i = hhashsucc(S->T, h, 0);
   while (i >= 0) {
-    uint x = hhashget(S->T, h, i);
+    ulong x = hhashget(S->T, h, i);
     if (S->cmp(k, x) == 0) {
       if (kill) hhashdel(S->T, h, i);
       return x;
@@ -43,9 +43,9 @@ static uint hunt(HHSet *S, uint h, uint k, int kill) {
   return 0;
 }
 
-uint hhsetget(HHSet *S, uint k) { return hunt(S, S->hash(k), k, 0); }
+ulong hhsetget(HHSet *S, ulong k) { return hunt(S, S->hash(k), k, 0); }
 
-uint hhsetdel(HHSet *S, uint k) { return hunt(S, S->hash(k), k, 1); }
+ulong hhsetdel(HHSet *S, ulong k) { return hunt(S, S->hash(k), k, 1); }
 
 static int resize(HHSet *R) {
   HHSet *S = hhsetnew(2 * R->T->n, R->hash, R->cmp);
@@ -60,9 +60,9 @@ static int resize(HHSet *R) {
   return 1;
 }
 
-int hhsetput(HHSet *S, uint x) {
+int hhsetput(HHSet *S, ulong x) {
   if (x == 0) return 0;
-  uint h = S->hash(x);
+  ulong h = S->hash(x);
   if (hunt(S, h, x, 0)) return 0;
   if (hhashput(S->T, h, x)) return 1;
   if (!resize(S)) return 0;
