+++ test/set/test1.c
@@ -3,10 +3,10 @@
 #include <hhset.h>
 #include "test.h"
 
-static uint fill(HHSet *S) {
-  uint m = 0;
-  for (uint i = 0; i < 128; ++i) {
-    uint x = rand() + 1;
+static ulong fill(HHSet *S) {
+  ulong m = 0;
+  for (ulong i = 0; i < 128; ++i) {
+    ulong x = rand() + 1;
     if (!hhsetget(S, x)) {
       test(hhsetput(S, x));
       ++m;
@@ -15,17 +15,17 @@ static uint fill(HHSet *S) {
   return m;
 }
 
-static uint rnd(HHSet *S) {
+static ulong rnd(HHSet *S) {
   HHash *T = S->T;
-  uint h = rand() % T->n;
+  ulong h = rand() % T->n;
   while (T->V[h] == 0) h = (h + 1) % T->n;
   return T->V[h];
 }
 
-static uint strip(HHSet *S) {
-  uint m = 0;
+static ulong strip(HHSet *S) {
+  ulong m = 0;
   for (; m < 64; ++m) {
-    uint x = rnd(S);
+    ulong x = rnd(S);
     hhsetdel(S, x);
     test(!hhsetget(S, x));
   }
@@ -33,9 +33,9 @@ static uint strip(HHSet *S) {
 }
 
 int main(void) {
-  uint n = 256;
+  ulong n = 256;
   HHSet *S = hhsetnew(n, NULL, NULL);
-  uint m = fill(S);
+  ulong m = fill(S);
   test(m == S->T->m);
   m -= strip(S);
   test(m == S->T->m);
